imageLoad takes integer coordinates

+++ examples/01 Physically Based Rendering/assets/visualize_voxel.vert
@@ -19,7 +19,7 @@ void main()
   vec3 c = vec3(gl_InstanceID % gridResolution, 
                 (gl_InstanceID % (gridResolution*gridResolution)) / gridResolution,
                 gl_InstanceID / (gridResolution*gridResolution));
-  voxelColor = imageLoad(voxelTexture, vec3(c));
+  voxelColor = imageLoad(voxelTexture, ivec3(c));
 
   float scale = float(gridSize) / gridResolution;
   float gridStartPosition = gridSize / 2.0 - scale / 2.0;
